requirements.txt
################################################################################
# requirements.txt
################################################################################

# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.30.6

# GCP
google-cloud-aiplatform==1.68.0
google-cloud-secret-manager==2.20.2
google-cloud-storage==2.18.2
vertexai==1.68.0

# Database
SQLAlchemy==2.0.35
pg8000==1.31.2          # Pure Python PostgreSQL driver（Cloud Run 向け）
cloud-sql-python-connector[pg8000]==1.12.1

# GitHub Integration (Step 8)
PyGithub==2.4.0

# HTTP
requests==2.32.3
httpx==0.27.2

# Data Processing
pandas==2.2.3

# Testing
pytest==8.3.3
pytest-cov==5.0.0
pytest-asyncio==0.24.0

# Utilities
pydantic==2.9.2
python-multipart==0.0.12  # FastAPI file upload
